Fix gnome-games crashing due to missing Steam registry key
Long story short the registry might not have the InstallPath key, so we
should default to using the ~/.steam/steam symlink as the base of all
operations.

Unfortunately I cannot get the Vala code to recompile due to interdependencies
between the Steam module and the Retro module which doesn't emit a .vapi,
even though the Retro module doesn't emit one, thus the gnome-games.vapi
file is actually invalid in this context.

This commit also updates gnome-games to the latest upstream version, which
by itself did not fix the issue.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>